more backup of WIP - idea of protocol methods returning OperationResult and having factory methods to different LRO types for convenience and streaming
